chore(test): Split up tests implementation into multiple classes
AI-assistant: Claude Code v2.1.126 (claude-sonnet-4-6)
